shadow test + removed spatial audio
added a shadow sprite to act as a visual for the actual hitbox while the bird was moved up to have its legs be the center of the shadow. I also turned the sounds into normal audio streams so that mixing them is easier and less unpredictable. However, I still have yet to fix the item pickup duplication glitch.